fix: move getting stripe profile into auth adaptor
